refactor: 메서드명 및 final 선언 수정
